Nudge by a customizable number of pixels
For those who want more then just nudging by 1px (arrow keys) or 10px
(shift + arrow keys). For example in a recent project things usually
are moved in multiples of 16px. This is let me set this, and it's a
huge timesaver.

Uses cmd-opt plus arrow keys to nudge the selection by n pixels, where
n is any number. An additional cmd-opt-0 (zero) is used to set the
number